api/requirements.txt
firebase-admin==7.3.0
google-genai==1.68.0
